split mongo-specific fields into cleancat.mongo
